wip: generate quote with pubkey in report data
* needs review
* needs cleanup

The public key is written to file outside the enclave. Could this be a
problem? What if the untrusted app changes that public key? Same
reasoning goes for the report data as if the public key that is copied
into the report data of a quote is tampered with then the public key is
no longer useful to verify messages or outputs from the enclave.

Also:

* Add .c files from sgx-ra-sample
* TODO: Keep the history, using git-filter-repo
* Add example of verifying the signed data
* Make nix-related changes and additions