WIP: seal the public key instead of writing pem
The public key is being sealed to prevent that an attacker would swap
the key with a different one, such that the public key of the attacker
would end up in the report data of a quote.

The original implementation writes the public key to a buffer from which
the key is written to file. The operation is done in "untrusted" mode,
meaning NOT in an enclave, and consequently it appears that an attacker,
could modify the content of the buffer and/or file.

If the above assumption is correct, then a different mechanism is
required in order to make sure that the public key that will be put into
the report data of a quote will not have been tampered with, modified or
changed for a different one.

By tampering with the public key an attacker could perform two types of
attacks: 1) impersonation 2) denial of service.

1) Impersonation attack: the attacker swap the public key with its own,
and signs different data such that it looks like the data is indeed from
the enclave program when it fact it is not.

2) Denial of service attack: the attacker simply corrupts the public key
such that it is not valid, as it will not match the private key and
consequently data that is indeed authentic, truly originating from the
legit enclave, will be recognized as having an invalid signature and
therefore the enclave will be "denied". Perhaps "denial of service" is
not an accurate naming. The idea is that an authentic enclave is
recognized as being invalid, when it actual is.